proxy: Rename control::discovery to control::destination
This is more consistent with the naming in the controller.

Additionally, the destination module is now a directory so that its
components may be split up.
