+++ proxy/src/bind.rs
@@ -13,7 +13,7 @@ use tower_h2;
 use tower_reconnect::Reconnect;
 
 use control;
-use control::discovery::Endpoint;
+use control::destination::Endpoint;
 use ctx;
 use telemetry::{self, sensor};
 use transparency::{self, HttpBody, h1};
@@ -243,7 +243,7 @@ impl<C, B> Bind<C, B> {
     }
 }
 
-impl<B> control::discovery::Bind for BindProtocol<Arc<ctx::Proxy>, B>
+impl<B> control::destination::Bind for BindProtocol<Arc<ctx::Proxy>, B>
 where
     B: tower_h2::Body + 'static,
 {
